add apron java bindings as lib files
