+++ src/main/scala/edu/colorado/plv/cuanto/numdomain/apronapi/ApronInterval.scala
@@ -1,4 +1,4 @@
-package edu.colorado.plv.cuanto.numdomain.apron
+package edu.colorado.plv.cuanto.numdomain.apronapi
 
 import apron.Interval
 import edu.colorado.plv.cuanto.numdomain.AbsInterval
