+++ src/main/scala/edu/colorado/plv/cuanto/numdomain/apronapi/ApronNonLinExpr.scala
@@ -1,4 +1,4 @@
-package edu.colorado.plv.cuanto.numdomain.apron
+package edu.colorado.plv.cuanto.numdomain.apronapi
 
 import apron.Texpr0Intern
 import edu.colorado.plv.cuanto.numdomain.AbsExpr
